requirements.txt
absl-py @ file:///croot/absl-py_1714140470852/work
aiohttp @ file:///croot/aiohttp_1715108769648/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
antlr4-python3-runtime==4.9.3
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout @ file:///croot/async-timeout_1703096998144/work
attrs @ file:///croot/attrs_1695717823297/work
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker @ file:///croot/blinker_1696539051175/work
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools @ file:///croot/cachetools_1713977085444/work
causal-learn==0.1.3.7
certifi==2024.2.2
cffi @ file:///croot/cffi_1714483155441/work
charset-normalizer==3.1.0
chex==0.1.7
clarabel==0.6.0
click @ file:///croot/click_1698129812380/work
comm==0.2.1
contourpy==1.1.1
cryptography @ file:///croot/cryptography_1714660666131/work
cvxpy==1.4.1
cycler==0.12.1
Cython==0.29.36
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
densratio==0.3.0
dm-tree==0.1.8
docker-pycreds==0.4.0
dowhy==0.10.1
ecos==2.0.12
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
fonttools==4.46.0
fqdn==1.5.1
frozenlist @ file:///croot/frozenlist_1698702560391/work
fsspec==2023.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth @ file:///croot/google-auth_1715111149956/work
google-auth-oauthlib @ file:///opt/conda/conda-bld/google-auth-oauthlib_1660687784486/work
graphviz==0.20.1
grpcio @ file:///croot/grpc-suite_1701982571719/work
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
hydra-core==1.3.2
idna==3.6
importlib-metadata @ file:///croot/importlib_metadata-suite_1704813515092/work
importlib-resources==6.1.1
ipykernel==6.29.3
ipython==8.12.3
isoduration==20.11.0
jax==0.4.13
jaxlib==0.4.13+cuda11.cudnn86
jaxopt==0.8.3
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.20
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.3
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.13.0
jupyter_server_terminals==0.5.2
jupyterlab==4.1.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
kiwisolver==1.4.5
Markdown @ file:///croot/markdown_1671541909495/work
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
mistune==3.0.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
ml-dtypes==0.2.0
mpmath==1.3.0
multidict @ file:///croot/multidict_1701096859099/work
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.0
notebook_shim==0.2.4
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.3.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.3.101
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-nvcc-cu12==12.3.107
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.3.107
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.3.101
nvidia-cudnn-cu11==8.9.6.50
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.12.1
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.5.4.101
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.2.0.103
nvidia-nvjitlink-cu12==12.3.101
oauthlib @ file:///croot/oauthlib_1679489621486/work
omegaconf==2.3.0
opt-einsum==3.3.0
optax==0.1.8
osqp==0.6.3
overrides==7.7.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pybind11==2.11.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydot==1.4.2
Pygments==2.17.2
PyJWT @ file:///croot/pyjwt_1715094751094/work
pyOpenSSL @ file:///croot/pyopenssl_1708380408460/work
pyparsing==3.1.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qdldl==0.1.7.post0
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scikit-learn==1.3.2
scipy==1.10.1
scs==3.2.4.post1
seaborn==0.13.2
Send2Trash==1.8.2
sentry-sdk==2.7.0
setproctitle==1.3.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tensorboard @ file:///home/builder/mesters/opt/envs/tensorflow/conda-bld/tensorboard_1682445826165/work/tensorboard-2.12.1-py3-none-any.whl
tensorboard-data-server @ file:///croot/tensorboard-data-server_1681498183723/work/tensorboard_data_server-0.7.0-py3-none-manylinux2014_x86_64.whl
tensorboard-plugin-wit @ file:///home/builder/tkoch/workspace/tensorflow/tensorboard-plugin-wit_1658918494740/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
terminado==0.18.0
threadpoolctl==3.3.0
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.1
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchvision==0.16.1+cu118
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
triton==2.1.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.2.1
wandb==0.17.3
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug @ file:///croot/werkzeug_1716533313107/work
yarl @ file:///croot/yarl_1701105127787/work
zipp==3.17.0
